+++ src/post.js
@@ -1,18 +1,5 @@
-const core = require("@actions/core"); // https://github.com/actions/toolkit/tree/main/packages/core
-const exec = require("@actions/exec"); // https://github.com/actions/toolkit/tree/main/packages/exec
-
-// read action inputs
-const input = {
-  post: core.getInput('post', {required: true}),
-  workingDirectory: core.getInput('working-directory'),
-};
+const {post} = require("./common");
 
 (async () => {
-  const command = input.post
-
-  if (command !== "") {
-    await exec.exec(command, [], {cwd: input.workingDirectory});
-  }
-})().catch(error => {
-  core.error(error.message)
-});
+  await post()
+})();